+++ src/bucket/bucket.spec.ts
@@ -1,9 +1,9 @@
-import { Hose } from './hose';
+import { Bucket } from './bucket';
 
-describe('Hose', () => {
+describe('Bucket', () => {
   it('should flatten itself', () => {
-    const c1 = new Hose([1, 2, 3,]);
-    const c2 = new Hose(c1);
+    const c1 = new Bucket([1, 2, 3,]);
+    const c2 = new Bucket(c1);
     for (const next of c2) {
       expect(typeof next).toEqual('number');
     }
@@ -12,7 +12,7 @@ describe('Hose', () => {
   describe('run(...)', () => {
     describe('should work on', () => {
       it('array', () => {
-        const pipeline = new Hose([1, 2, 3,]);
+        const pipeline = new Bucket([1, 2, 3,]);
         const called: number[] = [];
         pipeline.exhaust(item => { called.push(item); });
         expect(called[0]).toEqual(1);
@@ -20,7 +20,7 @@ describe('Hose', () => {
         expect(called[2]).toEqual(3);
       });
       it('set', () => {
-        const pipeline = new Hose(new Set([1, 2, 3,]));
+        const pipeline = new Bucket(new Set([1, 2, 3,]));
         const called: number[] = [];
         pipeline.exhaust(item => { called.push(item); });
         expect(called[0]).toEqual(1);
@@ -29,7 +29,7 @@ describe('Hose', () => {
       });
       it('map', () => {
         const map = new Map<number, number>([[1, 1,], [2, 2,], [3, 3,],]);
-        const pipeline = new Hose(map);
+        const pipeline = new Bucket(map);
         const called: [number, number][] = [];
         pipeline.exhaust(item => { called.push(item); });
         expect(called[0]).toEqual([1, 1,]);
@@ -37,7 +37,7 @@ describe('Hose', () => {
         expect(called[2]).toEqual([3, 3,]);
       });
       it('generator', () => {
-        const pipeline = new Hose(function * (): Iterable<number> {
+        const pipeline = new Bucket(function * (): Iterable<number> {
           yield 1; yield 2; yield 3;
         });
         const called: number[] = [];
@@ -52,7 +52,7 @@ describe('Hose', () => {
   describe('map(...)', () => {
     describe('should work on', () => {
       it('array', () => {
-        const pipeline = new Hose([1, 2, 3,]).map(n => n + 1);
+        const pipeline = new Bucket([1, 2, 3,]).map(n => n + 1);
         const called: number[] = [];
         pipeline.exhaust(item => { called.push(item); });
         expect(called[0]).toEqual(2);
@@ -60,7 +60,7 @@ describe('Hose', () => {
         expect(called[2]).toEqual(4);
       });
       it('set', () => {
-        const pipeline = new Hose(new Set([1, 2, 3,])).map(n => n + 1);
+        const pipeline = new Bucket(new Set([1, 2, 3,])).map(n => n + 1);
         const called: number[] = [];
         pipeline.exhaust(item => { called.push(item); });
         expect(called[0]).toEqual(2);
@@ -69,7 +69,7 @@ describe('Hose', () => {
       });
       it('map', () => {
         const map = new Map<number, number>([[1, 1,], [2, 2,], [3, 3,],]);
-        const pipeline = new Hose(map).map<[number, number]>(
+        const pipeline = new Bucket(map).map<[number, number]>(
           ([a, b,]) => ([a + 1, b + 1,]));
         const called: [number, number][] = [];
         pipeline.exhaust(item => { called.push(item); });
@@ -78,7 +78,7 @@ describe('Hose', () => {
         expect(called[2]).toEqual([4, 4,]);
       });
       it('generator', () => {
-        const pipeline = new Hose(function * (): Iterable<number> {
+        const pipeline = new Bucket(function * (): Iterable<number> {
           yield 1;
           yield 2;
           yield 3;
@@ -95,26 +95,26 @@ describe('Hose', () => {
   describe('filter(...)', () => {
     describe('should work on', () => {
       it('array', () => {
-        const pipeline = new Hose([1, 2, 3,]).filter((v) => v !== 2);
+        const pipeline = new Bucket([1, 2, 3,]).filter((v) => v !== 2);
         const called: number[] = [];
         pipeline.exhaust(item => { called.push(item); });
         expect(called[1]).toEqual(3);
       });
       it('set', () => {
-        const pipeline = new Hose(new Set([1, 2, 3,])).filter((v) => v !== 2);
+        const pipeline = new Bucket(new Set([1, 2, 3,])).filter((v) => v !== 2);
         const called: number[] = [];
         pipeline.exhaust(item => { called.push(item); });
         expect(called[1]).toEqual(3);
       });
       it('map', () => {
         const map = new Map<number, number>([[1, 1,], [2, 2,], [3, 3,],]);
-        const pipeline = new Hose(map).filter(([v,]) => v !== 2);
+        const pipeline = new Bucket(map).filter(([v,]) => v !== 2);
         const called: [number, number][] = [];
         pipeline.exhaust(item => { called.push(item); });
         expect(called[1]).toEqual([3, 3,]);
       });
       it('generator', () => {
-        const pipeline = new Hose(function * (): Iterable<number> {
+        const pipeline = new Bucket(function * (): Iterable<number> {
           yield 1;
           yield 2;
           yield 3;
@@ -129,26 +129,26 @@ describe('Hose', () => {
   describe('exclude(...)', () => {
     describe('should work on', () => {
       it('array', () => {
-        const pipeline = new Hose([1, 2, 3,]).exclude(1, 2);
+        const pipeline = new Bucket([1, 2, 3,]).exclude(1, 2);
         const called: number[] = [];
         pipeline.exhaust(item => { called.push(item); });
         expect(called[0]).toEqual(3);
       });
       it('set', () => {
-        const pipeline = new Hose(new Set([1, 2, 3,])).exclude(1, 3);
+        const pipeline = new Bucket(new Set([1, 2, 3,])).exclude(1, 3);
         const called: number[] = [];
         pipeline.exhaust(item => { called.push(item); });
         expect(called[0]).toEqual(2);
       });
       it('map', () => {
         const map = new Map<number, number>([[1, 1,], [2, 2,], [3, 3,],]);
-        const pipeline = new Hose(() => map.values()).exclude(3, 1);
+        const pipeline = new Bucket(() => map.values()).exclude(3, 1);
         const called: number[] = [];
         pipeline.exhaust(item => { called.push(item); });
         expect(called[0]).toEqual(2);
       });
       it('generator', () => {
-        const pipeline = new Hose(function * (): Iterable<number> {
+        const pipeline = new Bucket(function * (): Iterable<number> {
           yield 1;
           yield 2;
           yield 3;
@@ -163,26 +163,26 @@ describe('Hose', () => {
   describe('pick(...)', () => {
     describe('should work on', () => {
       it('array', () => {
-        const pipeline = new Hose([1, 2, 3,]).pick(1, 3);
+        const pipeline = new Bucket([1, 2, 3,]).pick(1, 3);
         const called: number[] = [];
         pipeline.exhaust(item => { called.push(item); });
         expect(called[1]).toEqual(3);
       });
       it('set', () => {
-        const pipeline = new Hose(new Set([1, 2, 3,])).pick(1, 3);
+        const pipeline = new Bucket(new Set([1, 2, 3,])).pick(1, 3);
         const called: number[] = [];
         pipeline.exhaust(item => { called.push(item); });
         expect(called[1]).toEqual(3);
       });
       it('map', () => {
         const map = new Map<number, number>([[1, 1,], [2, 2,], [3, 3,],]);
-        const pipeline = new Hose(map.values()).pick(3, 1);
+        const pipeline = new Bucket(map.values()).pick(3, 1);
         const called: number[] = [];
         pipeline.exhaust(item => { called.push(item); });
         expect(called[1]).toEqual(3);
       });
       it('generator', () => {
-        const pipeline = new Hose(function * (): Iterable<number> {
+        const pipeline = new Bucket(function * (): Iterable<number> {
           yield 1;
           yield 2;
           yield 3;