+++ src/dam/dam.spec.ts
@@ -1,44 +1,44 @@
-import { Lake } from './lake';
+import { Dam } from './dam';
 
-describe('Lake', () => {
+describe('Dam', () => {
   it('should flatten itself', () => {
-    const c1 = new Lake([1, 2, 3,]);
-    const c2 = new Lake(c1);
+    const c1 = new Dam([1, 2, 3,]);
+    const c2 = new Dam(c1);
     for (const next of c2) {
       expect(typeof next).toEqual('number');
     }
   });
 
-  describe('item(...)', () => {
+  describe('at(...)', () => {
     describe('should work on', () => {
       it('array', () => {
-        const pipeline = new Lake([1, 2, 3,]);
-        expect(pipeline.item(0).value).toEqual(1);
-        expect(pipeline.item(1).value).toEqual(2);
-        expect(pipeline.item(2).value).toEqual(3);
+        const pipeline = new Dam([1, 2, 3,]);
+        expect(pipeline.at(0).value).toEqual(1);
+        expect(pipeline.at(1).value).toEqual(2);
+        expect(pipeline.at(2).value).toEqual(3);
       });
       it('set', () => {
-        const pipeline = new Lake(new Set([1, 2, 3,]));
-        expect(pipeline.item(0).value).toEqual(1);
-        expect(pipeline.item(1).value).toEqual(2);
-        expect(pipeline.item(2).value).toEqual(3);
+        const pipeline = new Dam(new Set([1, 2, 3,]));
+        expect(pipeline.at(0).value).toEqual(1);
+        expect(pipeline.at(1).value).toEqual(2);
+        expect(pipeline.at(2).value).toEqual(3);
       });
       it('map', () => {
         const map = new Map<number, number>([
           [1, 1,], [2, 2,], [3, 3,],
         ]);
-        const pipeline = new Lake(map);
-        expect(pipeline.item(0).value).toEqual([1, 1,]);
-        expect(pipeline.item(1).value).toEqual([2, 2,]);
-        expect(pipeline.item(2).value).toEqual([3, 3,]);
+        const pipeline = new Dam(map);
+        expect(pipeline.at(0).value).toEqual([1, 1,]);
+        expect(pipeline.at(1).value).toEqual([2, 2,]);
+        expect(pipeline.at(2).value).toEqual([3, 3,]);
       });
       it('generator', () => {
-        const pipeline = new Lake(function * (): Iterable<number> {
+        const pipeline = new Dam(function * (): Iterable<number> {
           yield 1; yield 2; yield 3;
         });
-        expect(pipeline.item(0).value).toEqual(1);
-        expect(pipeline.item(1).value).toEqual(2);
-        expect(pipeline.item(2).value).toEqual(3);
+        expect(pipeline.at(0).value).toEqual(1);
+        expect(pipeline.at(1).value).toEqual(2);
+        expect(pipeline.at(2).value).toEqual(3);
       });
     });
   });
@@ -46,7 +46,7 @@ describe('Lake', () => {
   describe('run(...)', () => {
     describe('should work on', () => {
       it('array', () => {
-        const pipeline = new Lake([1, 2, 3,]);
+        const pipeline = new Dam([1, 2, 3,]);
         const called: number[] = [];
         pipeline.forEach(item => { called.push(item); });
         expect(called[0]).toEqual(1);
@@ -54,7 +54,7 @@ describe('Lake', () => {
         expect(called[2]).toEqual(3);
       });
       it('set', () => {
-        const pipeline = new Lake(new Set([1, 2, 3,]));
+        const pipeline = new Dam(new Set([1, 2, 3,]));
         const called: number[] = [];
         pipeline.forEach(item => { called.push(item); });
         expect(called[0]).toEqual(1);
@@ -63,7 +63,7 @@ describe('Lake', () => {
       });
       it('map', () => {
         const map = new Map<number, number>([[1, 1,], [2, 2,], [3, 3,],]);
-        const pipeline = new Lake(map);
+        const pipeline = new Dam(map);
         const called: [number, number][] = [];
         pipeline.forEach(item => { called.push(item); });
         expect(called[0]).toEqual([1, 1,]);
@@ -71,12 +71,12 @@ describe('Lake', () => {
         expect(called[2]).toEqual([3, 3,]);
       });
       it('generator', () => {
-        const pipeline = new Lake(function * (): Iterable<number> {
+        const pipeline = new Dam(function * (): Iterable<number> {
           yield 1; yield 2; yield 3;
         });
-        expect(pipeline.item(0).value).toEqual(1);
-        expect(pipeline.item(1).value).toEqual(2);
-        expect(pipeline.item(2).value).toEqual(3);
+        expect(pipeline.at(0).value).toEqual(1);
+        expect(pipeline.at(1).value).toEqual(2);
+        expect(pipeline.at(2).value).toEqual(3);
       });
     });
   });
@@ -84,34 +84,34 @@ describe('Lake', () => {
   describe('map(...)', () => {
     describe('should work on', () => {
       it('array', () => {
-        const pipeline = new Lake([1, 2, 3,]).map(n => n + 1);
-        expect(pipeline.item(0).value).toEqual(2);
-        expect(pipeline.item(1).value).toEqual(3);
-        expect(pipeline.item(2).value).toEqual(4);
+        const pipeline = new Dam([1, 2, 3,]).map(n => n + 1);
+        expect(pipeline.at(0).value).toEqual(2);
+        expect(pipeline.at(1).value).toEqual(3);
+        expect(pipeline.at(2).value).toEqual(4);
       });
       it('set', () => {
-        const pipeline = new Lake(new Set([1, 2, 3,])).map(n => n + 1);
-        expect(pipeline.item(0).value).toEqual(2);
-        expect(pipeline.item(1).value).toEqual(3);
-        expect(pipeline.item(2).value).toEqual(4);
+        const pipeline = new Dam(new Set([1, 2, 3,])).map(n => n + 1);
+        expect(pipeline.at(0).value).toEqual(2);
+        expect(pipeline.at(1).value).toEqual(3);
+        expect(pipeline.at(2).value).toEqual(4);
       });
       it('map', () => {
         const map = new Map<number, number>([[1, 1,], [2, 2,], [3, 3,],]);
-        const pipeline = new Lake(map).map<[number, number]>(
+        const pipeline = new Dam(map).map<[number, number]>(
           ([a, b,]) => ([a + 1, b + 1,]));
-        expect(pipeline.item(0).value).toEqual([2, 2,]);
-        expect(pipeline.item(1).value).toEqual([3, 3,]);
-        expect(pipeline.item(2).value).toEqual([4, 4,]);
+        expect(pipeline.at(0).value).toEqual([2, 2,]);
+        expect(pipeline.at(1).value).toEqual([3, 3,]);
+        expect(pipeline.at(2).value).toEqual([4, 4,]);
       });
       it('generator', () => {
-        const pipeline = new Lake(function * (): Iterable<number> {
+        const pipeline = new Dam(function * (): Iterable<number> {
           yield 1;
           yield 2;
           yield 3;
         }).map(n => n + 1);
-        expect(pipeline.item(0).value).toEqual(2);
-        expect(pipeline.item(1).value).toEqual(3);
-        expect(pipeline.item(2).value).toEqual(4);
+        expect(pipeline.at(0).value).toEqual(2);
+        expect(pipeline.at(1).value).toEqual(3);
+        expect(pipeline.at(2).value).toEqual(4);
       });
     });
   });
@@ -119,20 +119,20 @@ describe('Lake', () => {
   describe('first(...)', () => {
     describe('should work on', () => {
       it('array', () => {
-        const pipeline = new Lake([1, 2, 3,]);
+        const pipeline = new Dam([1, 2, 3,]);
         expect(pipeline.first().value).toEqual(1);
       });
       it('set', () => {
-        const pipeline = new Lake(new Set([1, 2, 3,]));
+        const pipeline = new Dam(new Set([1, 2, 3,]));
         expect(pipeline.first().value).toEqual(1);
       });
       it('map', () => {
         const map = new Map<number, number>([[1, 1,], [2, 2,], [3, 3,],]);
-        const pipeline = new Lake(map);
+        const pipeline = new Dam(map);
         expect(pipeline.first().value).toEqual([1, 1,]);
       });
       it('generator', () => {
-        const pipeline = new Lake(function * (): Iterable<number> {
+        const pipeline = new Dam(function * (): Iterable<number> {
           yield 1;
           yield 2;
           yield 3;
@@ -145,25 +145,25 @@ describe('Lake', () => {
   describe('filter(...)', () => {
     describe('should work on', () => {
       it('array', () => {
-        const pipeline = new Lake([1, 2, 3,]).filter((v) => v !== 2);
-        expect(pipeline.item(1).value).toEqual(3);
+        const pipeline = new Dam([1, 2, 3,]).filter((v) => v !== 2);
+        expect(pipeline.at(1).value).toEqual(3);
       });
       it('set', () => {
-        const pipeline = new Lake(new Set([1, 2, 3,])).filter((v) => v !== 2);
-        expect(pipeline.item(1).value).toEqual(3);
+        const pipeline = new Dam(new Set([1, 2, 3,])).filter((v) => v !== 2);
+        expect(pipeline.at(1).value).toEqual(3);
       });
       it('map', () => {
         const map = new Map<number, number>([[1, 1,], [2, 2,], [3, 3,],]);
-        const pipeline = new Lake(map).filter(([v,]) => v !== 2);
-        expect(pipeline.item(1).value).toEqual([3, 3,]);
+        const pipeline = new Dam(map).filter(([v,]) => v !== 2);
+        expect(pipeline.at(1).value).toEqual([3, 3,]);
       });
       it('generator', () => {
-        const pipeline = new Lake(function * (): Iterable<number> {
+        const pipeline = new Dam(function * (): Iterable<number> {
           yield 1;
           yield 2;
           yield 3;
         }).filter((v) => v !== 2);
-        expect(pipeline.item(1).value).toEqual(3);
+        expect(pipeline.at(1).value).toEqual(3);
       });
     });
   });
@@ -171,25 +171,25 @@ describe('Lake', () => {
   describe('exclude(...)', () => {
     describe('should work on', () => {
       it('array', () => {
-        const pipeline = new Lake([1, 2, 3,]).exclude(1, 2);
-        expect(pipeline.item(0).value).toEqual(3);
+        const pipeline = new Dam([1, 2, 3,]).exclude(1, 2);
+        expect(pipeline.at(0).value).toEqual(3);
       });
       it('set', () => {
-        const pipeline = new Lake(new Set([1, 2, 3,])).exclude(1, 3);
-        expect(pipeline.item(0).value).toEqual(2);
+        const pipeline = new Dam(new Set([1, 2, 3,])).exclude(1, 3);
+        expect(pipeline.at(0).value).toEqual(2);
       });
       it('map', () => {
         const map = new Map<number, number>([[1, 1,], [2, 2,], [3, 3,],]);
-        const pipeline = new Lake(() => map.values()).exclude(3, 1);
-        expect(pipeline.item(0).value).toEqual(2);
+        const pipeline = new Dam(() => map.values()).exclude(3, 1);
+        expect(pipeline.at(0).value).toEqual(2);
       });
       it('generator', () => {
-        const pipeline = new Lake(function * (): Iterable<number> {
+        const pipeline = new Dam(function * (): Iterable<number> {
           yield 1;
           yield 2;
           yield 3;
         }).exclude(2);
-        expect(pipeline.item(1).value).toEqual(3);
+        expect(pipeline.at(1).value).toEqual(3);
       });
     });
   });
@@ -197,25 +197,25 @@ describe('Lake', () => {
   describe('pick(...)', () => {
     describe('should work on', () => {
       it('array', () => {
-        const pipeline = new Lake([1, 2, 3,]).pick(1, 3);
-        expect(pipeline.item(1).value).toEqual(3);
+        const pipeline = new Dam([1, 2, 3,]).pick(1, 3);
+        expect(pipeline.at(1).value).toEqual(3);
       });
       it('set', () => {
-        const pipeline = new Lake(new Set([1, 2, 3,])).pick(1, 3);
-        expect(pipeline.item(1).value).toEqual(3);
+        const pipeline = new Dam(new Set([1, 2, 3,])).pick(1, 3);
+        expect(pipeline.at(1).value).toEqual(3);
       });
       it('map', () => {
         const map = new Map<number, number>([[1, 1,], [2, 2,], [3, 3,],]);
-        const pipeline = new Lake(map.values()).pick(3, 1);
-        expect(pipeline.item(1).value).toEqual(3);
+        const pipeline = new Dam(map.values()).pick(3, 1);
+        expect(pipeline.at(1).value).toEqual(3);
       });
       it('generator', () => {
-        const pipeline = new Lake(function * (): Iterable<number> {
+        const pipeline = new Dam(function * (): Iterable<number> {
           yield 1;
           yield 2;
           yield 3;
         }).pick(2);
-        expect(pipeline.item(0).value).toEqual(2);
+        expect(pipeline.at(0).value).toEqual(2);
       });
     });
   });
@@ -223,20 +223,20 @@ describe('Lake', () => {
   describe('getSize(...)', () => {
     describe('should map every item', () => {
       it('array', () => {
-        const pipeline = new Lake([1, 2, 3,]);
+        const pipeline = new Dam([1, 2, 3,]);
         expect(pipeline.getSize()).toEqual(3);
       });
       it('set', () => {
-        const pipeline = new Lake(new Set([1, 2, 3,]));
+        const pipeline = new Dam(new Set([1, 2, 3,]));
         expect(pipeline.getSize()).toEqual(3);
       });
       it('map', () => {
         const map = new Map<number, number>([[1, 1,], [2, 2,], [3, 3,],]);
-        const pipeline = new Lake(map);
+        const pipeline = new Dam(map);
         expect(pipeline.getSize()).toEqual(3);
       });
       it('generator', () => {
-        const pipeline = new Lake(function * (): Iterable<number> {
+        const pipeline = new Dam(function * (): Iterable<number> {
           yield 1;
           yield 2;
           yield 3;
@@ -250,20 +250,20 @@ describe('Lake', () => {
   describe('push(...)', () => {
     describe('should work on', () => {
       it('array', () => {
-        const pipeline = new Lake([1, 2, 3,]).push(4, 5);
-        expect(pipeline.item(3).value).toEqual(4);
-        expect(pipeline.item(4).value).toEqual(5);
+        const pipeline = new Dam([1, 2, 3,]).push(4, 5);
+        expect(pipeline.at(3).value).toEqual(4);
+        expect(pipeline.at(4).value).toEqual(5);
       });
       it('set', () => {
-        const pipeline = new Lake(new Set([1, 2, 3,])).push(4, 5);
-        expect(pipeline.item(3).value).toEqual(4);
-        expect(pipeline.item(4).value).toEqual(5);
+        const pipeline = new Dam(new Set([1, 2, 3,])).push(4, 5);
+        expect(pipeline.at(3).value).toEqual(4);
+        expect(pipeline.at(4).value).toEqual(5);
       });
       it('map', () => {
         const map = new Map<number, number>([[1, 1,], [2, 2,], [3, 3,],]);
-        const pipeline = new Lake(() => map.values()).push(4, 5);
-        expect(pipeline.item(3).value).toEqual(4);
-        expect(pipeline.item(4).value).toEqual(5);
+        const pipeline = new Dam(() => map.values()).push(4, 5);
+        expect(pipeline.at(3).value).toEqual(4);
+        expect(pipeline.at(4).value).toEqual(5);
       });
       it('generator', () => {
         const generator = function * (): Iterable<number> {
@@ -271,9 +271,9 @@ describe('Lake', () => {
           yield 2;
           yield 3;
         };
-        const pipeline = new Lake(generator).push(4, 5);
-        expect(pipeline.item(3).value).toEqual(4);
-        expect(pipeline.item(4).value).toEqual(5);
+        const pipeline = new Dam(generator).push(4, 5);
+        expect(pipeline.at(3).value).toEqual(4);
+        expect(pipeline.at(4).value).toEqual(5);
       });
     });
   });
@@ -282,20 +282,20 @@ describe('Lake', () => {
   describe('concat(...)', () => {
     describe('should work on', () => {
       it('array', () => {
-        const pipeline = new Lake([1, 2, 3,]).concat([4, 5,]);
-        expect(pipeline.item(3).value).toEqual(4);
-        expect(pipeline.item(4).value).toEqual(5);
+        const pipeline = new Dam([1, 2, 3,]).concat([4, 5,]);
+        expect(pipeline.at(3).value).toEqual(4);
+        expect(pipeline.at(4).value).toEqual(5);
       });
       it('set', () => {
-        const pipeline = new Lake(new Set([1, 2, 3,])).concat([4, 5,]);
-        expect(pipeline.item(3).value).toEqual(4);
-        expect(pipeline.item(4).value).toEqual(5);
+        const pipeline = new Dam(new Set([1, 2, 3,])).concat([4, 5,]);
+        expect(pipeline.at(3).value).toEqual(4);
+        expect(pipeline.at(4).value).toEqual(5);
       });
       it('map', () => {
         const map = new Map<number, number>([[1, 1,], [2, 2,], [3, 3,],]);
-        const pipeline = new Lake(() => map.values()).concat([4, 5,]);
-        expect(pipeline.item(3).value).toEqual(4);
-        expect(pipeline.item(4).value).toEqual(5);
+        const pipeline = new Dam(() => map.values()).concat([4, 5,]);
+        expect(pipeline.at(3).value).toEqual(4);
+        expect(pipeline.at(4).value).toEqual(5);
       });
       it('generator', () => {
         const generator = function * (): Iterable<number> {
@@ -303,9 +303,9 @@ describe('Lake', () => {
           yield 2;
           yield 3;
         };
-        const pipeline = new Lake(generator).concat([4, 5,]);
-        expect(pipeline.item(3).value).toEqual(4);
-        expect(pipeline.item(4).value).toEqual(5);
+        const pipeline = new Dam(generator).concat([4, 5,]);
+        expect(pipeline.at(3).value).toEqual(4);
+        expect(pipeline.at(4).value).toEqual(5);
       });
     });
   });
@@ -314,7 +314,7 @@ describe('Lake', () => {
   describe('sort(...)', () => {
     describe('should work on', () => {
       it('array', () => {
-        const pipeline = new Lake([1, 2, 3,]).sort(-1);
+        const pipeline = new Dam([1, 2, 3,]).sort(-1);
         expect(pipeline.sort(-1).toArray()).toEqual([3, 2, 1,]);
         expect(pipeline.sort(1).toArray()).toEqual([1, 2, 3,]);
         expect(pipeline.sort('desc').toArray()).toEqual([3, 2, 1,]);
@@ -325,7 +325,7 @@ describe('Lake', () => {
         expect(pipeline.sort('1').toArray()).toEqual([1, 2, 3,]);
       });
       it('set', () => {
-        const pipeline = new Lake(new Set([1, 2, 3,]));
+        const pipeline = new Dam(new Set([1, 2, 3,]));
         expect(pipeline.sort(-1).toArray()).toEqual([3, 2, 1,]);
         expect(pipeline.sort(1).toArray()).toEqual([1, 2, 3,]);
         expect(pipeline.sort('desc').toArray()).toEqual([3, 2, 1,]);
@@ -337,7 +337,7 @@ describe('Lake', () => {
       });
       it('map', () => {
         const map = new Map<number, number>([[1, 1,], [2, 2,], [3, 3,],]);
-        const pipeline = new Lake(() => map.values());
+        const pipeline = new Dam(() => map.values());
         expect(pipeline.sort(-1).toArray()).toEqual([3, 2, 1,]);
         expect(pipeline.sort(1).toArray()).toEqual([1, 2, 3,]);
         expect(pipeline.sort('desc').toArray()).toEqual([3, 2, 1,]);
@@ -353,7 +353,7 @@ describe('Lake', () => {
           yield 2;
           yield 3;
         };
-        const pipeline = new Lake(generator);
+        const pipeline = new Dam(generator);
         expect(pipeline.sort(-1).toArray()).toEqual([3, 2, 1,]);
         expect(pipeline.sort(1).toArray()).toEqual([1, 2, 3,]);
         expect(pipeline.sort('desc').toArray()).toEqual([3, 2, 1,]);
@@ -369,16 +369,16 @@ describe('Lake', () => {
   describe('toArray(...)', () => {
     describe('should work on', () => {
       it('array', () => {
-        const pipeline = new Lake([1, 2, 3,]);
+        const pipeline = new Dam([1, 2, 3,]);
         expect(pipeline.toArray()).toEqual([1, 2, 3,]);
       });
       it('set', () => {
-        const pipeline = new Lake(new Set([1, 2, 3,]));
+        const pipeline = new Dam(new Set([1, 2, 3,]));
         expect(pipeline.toArray()).toEqual([1, 2, 3,]);
       });
       it('map', () => {
         const map = new Map<number, number>([[1, 1,], [2, 2,], [3, 3,],]);
-        const pipeline = new Lake(map);
+        const pipeline = new Dam(map);
         expect(pipeline.toArray()).toEqual([[1, 1,], [2, 2,], [3, 3,],]);
       });
       it('generator', () => {
@@ -387,7 +387,7 @@ describe('Lake', () => {
           yield 2;
           yield 3;
         };
-        const pipeline = new Lake(generator);
+        const pipeline = new Dam(generator);
         expect(pipeline.toArray()).toEqual([1, 2, 3,]);
       });
     });
@@ -396,16 +396,16 @@ describe('Lake', () => {
   describe('toSet(...)', () => {
     describe('should work on', () => {
       it('array', () => {
-        const pipeline = new Lake([1, 2, 3,]);
+        const pipeline = new Dam([1, 2, 3,]);
         expect(pipeline.toSet()).toEqual(new Set([1, 2, 3,]));
       });
       it('set', () => {
-        const pipeline = new Lake(new Set([1, 2, 3,]));
+        const pipeline = new Dam(new Set([1, 2, 3,]));
         expect(pipeline.toSet()).toEqual(new Set([1, 2, 3,]));
       });
       it('map', () => {
         const map = new Map<number, number>([[1, 1,], [2, 2,], [3, 3,],]);
-        const pipeline = new Lake(map);
+        const pipeline = new Dam(map);
         expect(pipeline.toSet()).toEqual(new Set([[1, 1,], [2, 2,], [3, 3,],]));
       });
       it('generator', () => {
@@ -414,7 +414,7 @@ describe('Lake', () => {
           yield 2;
           yield 3;
         };
-        const pipeline = new Lake(generator);
+        const pipeline = new Dam(generator);
         expect(pipeline.toSet()).toEqual(new Set([1, 2, 3,]));
       });
     });