+++ src/dam/dam.ts
@@ -1,31 +1,31 @@
 import { Maybe } from '@nkp/maybe';
-import { HKT, Kind, URIs } from '../HKT';
+import { GetURI, HKT, Kind, URIs } from '../HKT';
 import { Pipeline } from '../pipeline/pipeline';
 import { Pipelineable, unpipeline } from '../utils/types';
 import { $TODO } from '../utils/utility-types';
 
 // declare URI
-export const LakeURI = 'Lake';
-export type LakeURI = typeof LakeURI;
+export const DamURI = 'Dam';
+export type DamURI = typeof DamURI;
 
 // register for usage as higher kind type
 declare module '../HKT' {
   interface URIToKind<A> {
-    readonly [LakeURI]: Lake<A>;
+    readonly [DamURI]: Dam<A>;
   }
 }
 
 /**
- * Iterable Lake
+ * Iterable Dam
  *
- * Items rest calmy in the lake until the foodgates open.
+ * Items rest calmy in the dam until the foodgates open.
  *
  *  - memory: light
  *  - cpu: heavy
  *
- * Items are created on-demand when an endpoint function
- * like toArray(), toSet(), item(...), or getSize(...) are called.
- * Items are otherwise not stored in memory
+ * Items are created on-demand when an endpoint function like toArray(),
+ * toSet(), item(...), or getSize(...) are called.
+ * Items are otherwise not stored in the Dam's memory
  *
  * Do ! NOT ! use directly with exhaustive iterators like:
  *  - Array.prototype.values
@@ -42,17 +42,17 @@ declare module '../HKT' {
  * Bad:
  * ```ts
  * const map = new Map([[1, {}, [2, {}], [3, {}]])
- * new Lake(map.values());
+ * new Dam(map.values());
  * ```
  *
  * Good:
  * ```ts
  * const map = new Map([[1, {}, [2, {}], [3, {}]]);
- * new Lake(() => map.values())
+ * new Dam(() => map.values())
  * ```
  */
-export class Lake<T> extends Pipeline<T> implements HKT<LakeURI, T> {
-  public override readonly URI = LakeURI;
+export class Dam<T> extends Pipeline<T> implements HKT<DamURI, T> {
+  public override readonly URI = DamURI;
 
   constructor(protected readonly root: Pipelineable<T>) {
     super();
@@ -65,8 +65,8 @@ export class Lake<T> extends Pipeline<T> implements HKT<LakeURI, T> {
    * @param
    * @param T
    */
-  static record<T>(iterable: Iterable<T>): Lake<T> {
-    return new Lake<T>(Array.from(iterable));
+  static record<T>(iterable: Iterable<T>): Dam<T> {
+    return new Dam<T>(Array.from(iterable));
   }
 
   /**
@@ -77,11 +77,11 @@ export class Lake<T> extends Pipeline<T> implements HKT<LakeURI, T> {
   }
 
   /**
-   * Collect the lakes contents
+   * Collect the Dam's contents
    *
    * Runs all transformations and rebases to an array of the resulting items
    */
-  rebase<S extends URIs>(this: Kind<S, T>): Kind<S, T> {
+  rebase<H1 extends URIs>(this: Kind<H1, T>): Kind<H1, T> {
     return new (this.constructor as $TODO)(this.toArray());
   }
 
@@ -90,10 +90,10 @@ export class Lake<T> extends Pipeline<T> implements HKT<LakeURI, T> {
    *
    * @param callbackfn
    */
-  forEach<S extends URIs>(
-    this: Kind<S, T>,
+  forEach<H1 extends URIs = GetURI<this>>(
+    this: Kind<H1, T>,
     callbackfn: (item: T, i: number, stop: () => void) => unknown,
-  ): Kind<S, T> {
+  ): Kind<H1, T> {
     let stopped = false;
     const stop = () => { stopped = true; };
     let i = 0;
@@ -110,17 +110,33 @@ export class Lake<T> extends Pipeline<T> implements HKT<LakeURI, T> {
    *
    * @returns
    */
-  first(): Maybe<T> {
-    return this.item(0);
+  first<H1 extends URIs = GetURI<this>>(this: Kind<H1, T>): Maybe<T> {
+    return (this as Dam<T>).at(0);
+  }
+
+  /**
+   * Find an item in the iterable
+   */
+  find<H1 extends URIs = GetURI<this>>(
+    this: Kind<H1, T>,
+    callbackfn: (value: T, currentIndex: number) => boolean,
+  ): Maybe<T> {
+    let i = 0;
+    for (const item of this) {
+      if (callbackfn(item, i)) return Maybe.some(item);
+      i += 1;
+    }
+    return Maybe.none;
   }
 
   /**
    * Get the current size of the pipeline if run
    */
-  getSize(): number {
-    if (Array.isArray(this.root)) return this.root.length;
-    if (this.root instanceof Set) return this.root.size;
-    if (this.root instanceof Map) return this.root.size;
+  getSize<H1 extends URIs = GetURI<this>>(this: Kind<H1, T>): number {
+    const self = this as Dam<T>;
+    if (Array.isArray(self.root)) return self.root.length;
+    if (self.root instanceof Set) return self.root.size;
+    if (self.root instanceof Map) return self.root.size;
     let i = 0;
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
     for (const _ of this) { i += 1; }
@@ -139,7 +155,7 @@ export class Lake<T> extends Pipeline<T> implements HKT<LakeURI, T> {
    * @param index
    * @returns
    */
-  item(index: number): Maybe<T> {
+  at(index: number): Maybe<T> {
     // shortcut for arrays
     if (Array.isArray(this.root)) {
       if (index >= 0) {
@@ -162,7 +178,7 @@ export class Lake<T> extends Pipeline<T> implements HKT<LakeURI, T> {
     } else {
       // i is negative
       // collect as array and reverse index
-      const arr = (this as Kind<LakeURI, T>).toArray();
+      const arr = (this as Kind<DamURI, T>).toArray();
       if (-index > arr.length) return Maybe.none;
       return Maybe.some(arr[arr.length - index]!);
     }